refactor: mock 데이터 생성 코드 분리
[#851]
