+++ frontend/.gitignore
@@ -4,4 +4,3 @@
 .husky
 yarn-error.log
 /storybook-static/
-/src/mocks/assets/
