--- frontend/src/mocks/json.d.ts
@@ -1,1 +1,0 @@
-declare module '*.json';
